+++ plugins/catalog-backend-module-github/src/providers/GithubEntityProvider.ts
@@ -163,7 +163,10 @@ export class GithubEntityProvider implements EntityProvider, EventSubscriber {
           try {
             await this.refresh(logger);
           } catch (error) {
-            logger.error(`${this.getProviderName()} refresh failed`, error);
+            logger.error(
+              `${this.getProviderName()} refresh failed, ${error}`,
+              error,
+            );
           }
         },
       });