+++ plugins/catalog-backend-module-github/src/providers/GithubOrgEntityProvider.ts
@@ -584,7 +584,10 @@ export class GithubOrgEntityProvider
           try {
             await this.read({ logger });
           } catch (error) {
-            logger.error(`${this.getProviderName()} refresh failed`, error);
+            logger.error(
+              `${this.getProviderName()} refresh failed, ${error}`,
+              error,
+            );
           }
         },
       });